pyproject.toml
[project]
name = "astro-reason"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kagglehub[pandas-datasets]>=0.3.13",
    "matplotlib>=3.10.7",
    "mcp>=1.19.0",
    "openai>=2.8.1",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sgp4>=2.25",
    "skyfield>=1.53",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "numpy>=2.3.5",
]

[tool.pytest.ini_options]
pythonpath = ["src", "satnet"]
testpaths = ["tests"]

[dependency-groups]
satnet = [
    "gym>=0.26.2",
    "gymnasium>=1.2.3",
    "numba>=0.63.1",
    "ray>=2.53.0",
    "sortedcontainers>=2.4.0",
]
